+++ src/server/wesnothd/game.cpp
@@ -391,7 +391,7 @@ void game::reset_sides()
 	side_controllers_.clear();
 	sides_.clear();
 
-	nsides_ = get_sides_list().size();
+	nsides_ = static_cast<int>(get_sides_list().size());
 
 	side_controllers_.resize(nsides_);
 	sides_.resize(nsides_);
@@ -638,7 +638,7 @@ bool game::describe_slots()
 	}
 
 	int available_slots = 0;
-	int num_sides = get_sides_list().size();
+	int num_sides = static_cast<int>(get_sides_list().size());
 	int i = 0;
 
 	for(const simple_wml::node* side : get_sides_list()) {
@@ -985,7 +985,7 @@ bool game::process_turn(simple_wml::document& data, player_iterator user)
 			LOG_GAME << msg.str() << " (game id: " << id_ << ", " << db_id_ << ")";
 			send_and_record_server_message(msg.str());
 
-			marked.push_back(index - marked.size());
+			marked.push_back(index - static_cast<int>(marked.size()));
 		} else if((*command).child("speak")) {
 			simple_wml::node& speak = *(*command).child("speak");
 			if(!speak["to_sides"].empty() || is_muted_observer(user)) {
@@ -1345,7 +1345,7 @@ void game::update_turn_data()
 		turns_cfg = &description_->add_child("turn_data");
 	}
 
-	turns_cfg->set_attr_int("current", current_turn());
+	turns_cfg->set_attr_int("current", static_cast<int>(current_turn()));
 	turns_cfg->set_attr_int("max", num_turns_);
 }
